Create README.md
What I'm finding when I try to do run npm in my npm command on my console application I'm getting request for money.  This was not the case a few years back when I was using node and npm files.  I need 8+ sponsored apps, of which the first one wanted $30 per month.  Why  has this monetization taken effect; usually this is done on a voluntary basis.  I'm just using this stuff to brush u[ on my skills for getting some side gigs for fun.  I was completely satisfied with the older versions of the code where I didn't need to pay.  It seems that these sponsors improved the code of there own free will, why are they now dissatisfied???